+++ view/Usuario/list_usuario.php
@@ -4,6 +4,14 @@ To change this license header, choose License Headers in Project Properties.
 To change this template file, choose Tools | Templates
 and open the template in the editor.
 -->
+
+<?php
+session_start(); 
+if((substr_compare($_SESSION['permissao']['usuario'], '0', 0, 1)) == 0) {
+    header("Location: ../Erro/permissao.php");
+}
+?>
+
 <html>
     <head>
         <!-- Required meta tags -->
@@ -35,7 +43,7 @@ and open the template in the editor.
                         </a>
 
                         <div class="dropdown-menu" aria-labelledby="dropdownMenuLink">
-                            <a class="dropdown-item" href="#"><?php session_start(); 
+                            <a class="dropdown-item" href="#"><?php
                                                                     if(isset($_SESSION['usuario'])) {
                                                                         echo 'Usuário: '. $_SESSION['usuario'];
                                                                     } else {
@@ -47,10 +55,13 @@ and open the template in the editor.
                     </div>
                 </div>
               </div>
-                </br>
                 <div style="text-align: right;">
                     <p>
-                        <a href="create_usuario.php" class="btn btn-outline-success">Adicionar</a>
+                        <?php
+                        if((substr_compare($_SESSION['permissao']['usuario'], '1', 1,1)) == 0) {
+                            echo '<a href="create_usuario.php" class="btn btn-outline-success">Adicionar</a>';
+                        }
+                        ?>
                     </p>
                 </div>
                         <table style="width: 100%;" id="table" class="table table-striped" data-toggle="table" data-search="true" data-pagination="true"
@@ -59,8 +70,14 @@ and open the template in the editor.
                             <tr>
                                 <th scope="col" data-field="id" data-sortable="true">Id</th>
                                 <th scope="col" data-field="usuario" data-sortable="true">Usuário</th>
-                                <th scope="col">Atualizar</th>
-                                <th scope="col">Excluir</th>
+                                <?php
+                                if ((substr_compare($_SESSION['permissao']['usuario'], '1', 2, 1)) == 0) {
+                                    echo '<th scope="col">Atualizar</th>';
+                                }
+                                if ((substr_compare($_SESSION['permissao']['usuario'], '1', 3, 1)) == 0) {
+                                    echo '<th scope="col">Excluir</th>';
+                                }
+                                ?>
                             </tr>
                         </thead>
                         <tbody>
@@ -75,9 +92,13 @@ and open the template in the editor.
                                 echo '<td>'. $row['id'] . '</td>';
                                 echo '<td>'. $row['usuario'] . '</td>';
                                 echo ' ';
-                                echo '<td width="80"><a class="btn btn-outline-warning btn-sm" href="update_usuario.php?id='.$row['id'].'">Atualizar</a></td>';
+                                if ((substr_compare($_SESSION['permissao']['usuario'], '1', 2, 1)) == 0) {
+                                    echo '<td width="80"><a class="btn btn-outline-warning btn-sm" href="update_usuario.php?id='.$row['id'].'">Atualizar</a></td>';
+                                }
                                 echo ' ';
-                                echo '<td width="80"><a class="btn btn-outline-danger btn-sm" href="delete_usuario.php?id='.$row['id'].'"> Excluir </a></td>';
+                                if ((substr_compare($_SESSION['permissao']['usuario'], '1', 3, 1)) == 0) {
+                                    echo '<td width="80"><a class="btn btn-outline-danger btn-sm" href="delete_usuario.php?id='.$row['id'].'"> Excluir </a></td>';
+                                }
                                 echo ' ';
                                 echo '</tr>';
                             }
